Added AdminCheck class and implemented logic behidn displaying admin functionalities
