+++ Android/app/src/main/java/com/example/coursecontrol/addNewCourse/ProgramDisplayActivity.kt
@@ -1,11 +1,13 @@
 package com.example.coursecontrol.addNewCourse
 
+import AdminChecker
 import UserDataAdapter
 import android.app.AlertDialog
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
+import android.view.View
 import android.widget.Button
 import android.widget.EditText
 import androidx.activity.viewModels
@@ -31,6 +33,7 @@ import kotlinx.coroutines.launch
 class ProgramDisplayActivity : AppCompatActivity() {
     private val viewModel: ProgramViewModel by viewModels()
     private val viewProgramModel: CourseAddNewProgram by viewModels()
+    val adminChecker = AdminChecker()
     private lateinit var sessionManager: SessionManager
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -42,8 +45,12 @@ class ProgramDisplayActivity : AppCompatActivity() {
         recyclerView.layoutManager = LinearLayoutManager(this)
 
         val addNewProgramButton: Button = findViewById(R.id.btnAddNewProgram)
-        addNewProgramButton.setOnClickListener {
+        val isAdmin = adminChecker.isAdmin()
 
+        if (isAdmin) {
+            addNewProgramButton.visibility = View.GONE
+        }
+        addNewProgramButton.setOnClickListener {
             showNewProgramDialog()
         }
 
@@ -76,6 +83,7 @@ class ProgramDisplayActivity : AppCompatActivity() {
         lifecycleScope.launch {
             try {
                 val sessionToken = sessionManager.getSessionToken()
+                adminChecker.checkAdmin(sessionToken)
                 if (sessionToken != null) {
                     viewModel.makeApiCall()
                 } else {
